+++ campari/run_one_object.py
@@ -18,7 +18,7 @@ from roman_imsim.utils import roman_utils
 from campari.data_construction import construct_images, prep_data_for_fit
 from campari.model_building import construct_static_scene, construct_transient_scene, generate_guess, make_grid
 from campari.simulation import simulate_images
-from campari.utils import banner, get_weights
+from campari.utils import banner, get_weights, campari_lightcurve_model
 from snpit_utils.config import Config
 from snpit_utils.logger import SNLogger
 
@@ -93,7 +93,7 @@ def run_one_object(diaobj=None, object_type=None, image_list=None,
     else:
         # Simulate the images of the SN and galaxy.
         banner("Simulating Images")
-        sim_lc, util_ref, image_list, cutout_image_list, sim_galra, sim_galdec, galaxy_images, noise_maps = \
+        simulated_lightcurve, util_ref = \
             simulate_images(image_list, diaobj,
                             sim_galaxy_scale, sim_galaxy_offset,
                             do_xshift, do_rotation, noise=noise,
@@ -104,6 +104,13 @@ def run_one_object(diaobj=None, object_type=None, image_list=None,
                             source_phot_ops=source_phot_ops,
                             mismatch_seds=mismatch_seds, base_pointing=base_pointing,
                             base_sca=base_sca)
+        sim_lc = simulated_lightcurve.sim_lc
+        image_list = simulated_lightcurve.image_list
+        cutout_image_list = simulated_lightcurve.cutout_image_list
+        galaxy_images = simulated_lightcurve.galaxy_images
+        noise_maps = simulated_lightcurve.noise_maps
+        sim_galra = simulated_lightcurve.galra
+        sim_galdec = simulated_lightcurve.galdec
         object_type = "SN"
 
     # Build the background grid
@@ -363,81 +370,3 @@ def run_one_object(diaobj=None, object_type=None, image_list=None,
         )
 
     return lightcurve_model
-
-
-class campari_lightcurve_model:
-    """This class holds the output of the Campari pipeline for a single SNID."""
-
-    def __init__(
-        self,
-        diaobj=None,
-        flux=None,
-        sigma_flux=None,
-        images=None,
-        model_images=None,
-        image_list=None,
-        cutout_image_list=None,
-        ra_grid=None,
-        dec_grid=None,
-        wgt_matrix=None,
-        confusion_metric=None,
-        best_fit_model_values=None,
-        sim_lc=None,
-        galaxy_images=None,
-        noise_maps=None,
-        galaxy_only_model_images=None,
-    ):
-        """Initialize the Campari lightcurve model with the SNID and its properties.
-        Parameters
-        ----------
-        diaobj: snappl.diaobject.DiaObject
-            The DiaObject representing the transient.
-        flux : np.ndarray
-            The flux values for the lightcurve.
-        sigma_flux : np.ndarray
-            The uncertainties in the flux values.
-        images : np.ndarray
-            The image data used in the lightcurve analysis.
-        model_images : np.ndarray
-            The model images generated by Campari.
-        image_list : list of snappl.image.Image objects
-            list of images used in the lightcurve analysis.
-        cutout_image_list : list of snappl.image.Image objects
-            list of images used in the lightcurve analysis that have been cutout at transient location.
-        ra_grid : np.ndarray
-            The RA coordinates of the points used to construct the background model.
-        dec_grid : np.ndarray
-            The Dec coordinates of the points used to construct the background model.
-        wgt_matrix : np.ndarray
-            The weight matrix used in the lightcurve analysis.
-        confusion_metric : np.ndarray
-            The confusion metric for the images. Currently defined as the dot product of PSF rendered
-            at the location of the transient and an image of the background galaxy. This is analogous to
-            local surface brightness, so it is possible this will be replaced with local surface brightness
-            in the future.
-        best_fit_model_values : np.ndarray
-            The best fit model values for the lightcurve. The last n values,
-            where n is the number of images considered a transient detection,
-            are the flux values for the transient. All other values are the
-            flux values assigned to the points that make up the background model.
-        cutout_wcs_list : list
-            List of WCS objects for the cutouts used in the lightcurve analysis.
-        sim_lc : pd.DataFrame
-            The simulated lightcurve data, if applicable.
-        """
-        self.diaobj = diaobj
-        self.flux = flux
-        self.sigma_flux = sigma_flux
-        self.images = images
-        self.model_images = model_images
-        self.image_list = image_list
-        self.cutout_image_list = cutout_image_list
-        self.ra_grid = ra_grid
-        self.dec_grid = dec_grid
-        self.wgt_matrix = wgt_matrix
-        self.confusion_metric = confusion_metric
-        self.best_fit_model_values = best_fit_model_values
-        self.sim_lc = sim_lc
-        self.galaxy_images = galaxy_images
-        self.noise_maps = noise_maps
-        self.galaxy_only_model_images = galaxy_only_model_images
