+++ campari/tests/test_campari.py
@@ -44,8 +44,11 @@ from campari.model_building import (
     make_regular_grid,
 )
 from campari.plotting import plot_lc
-from campari.run_one_object import campari_lightcurve_model
-from campari.utils import calc_mag_and_err, calculate_background_level, get_weights, make_sim_param_grid
+from campari.utils import (calc_mag_and_err,
+                           calculate_background_level,
+                           get_weights,
+                           make_sim_param_grid,
+                           campari_lightcurve_model)
 import snappl
 from snappl.diaobject import DiaObject
 from snappl.image import ManualFITSImage
