+++ campari/tests/testdata/partial_overlap.txt
@@ -1,0 +1,1 @@
+5934, 3
