+++ campari/utils.py
@@ -20,6 +20,92 @@ warnings.simplefilter("ignore", category=AstropyWarning)
 warnings.filterwarnings("ignore", category=ErfaWarning)
 
 
+class campari_lightcurve_model:
+    """This class holds the output of the Campari pipeline for a single SNID."""
+
+    def __init__(
+        self,
+        diaobj=None,
+        flux=None,
+        sigma_flux=None,
+        images=None,
+        model_images=None,
+        image_list=None,
+        cutout_image_list=None,
+        ra_grid=None,
+        dec_grid=None,
+        wgt_matrix=None,
+        confusion_metric=None,
+        best_fit_model_values=None,
+        sim_lc=None,
+        galaxy_images=None,
+        noise_maps=None,
+        galaxy_only_model_images=None,
+        galra=None,
+        galdec=None,
+    ):
+        """Initialize the Campari lightcurve model with the SNID and its properties.
+        Parameters
+        ----------
+        diaobj: snappl.diaobject.DiaObject
+            The DiaObject representing the transient.
+        flux : np.ndarray
+            The flux values for the lightcurve.
+        sigma_flux : np.ndarray
+            The uncertainties in the flux values.
+        images : np.ndarray
+            The image data used in the lightcurve analysis.
+        model_images : np.ndarray
+            The model images generated by Campari.
+        image_list : list of snappl.image.Image objects
+            list of images used in the lightcurve analysis.
+        cutout_image_list : list of snappl.image.Image objects
+            list of images used in the lightcurve analysis that have been cutout at transient location.
+        ra_grid : np.ndarray
+            The RA coordinates of the points used to construct the background model.
+        dec_grid : np.ndarray
+            The Dec coordinates of the points used to construct the background model.
+        wgt_matrix : np.ndarray
+            The weight matrix used in the lightcurve analysis.
+        confusion_metric : np.ndarray
+            The confusion metric for the images. Currently defined as the dot product of PSF rendered
+            at the location of the transient and an image of the background galaxy. This is analogous to
+            local surface brightness, so it is possible this will be replaced with local surface brightness
+            in the future.
+        best_fit_model_values : np.ndarray
+            The best fit model values for the lightcurve. The last n values,
+            where n is the number of images considered a transient detection,
+            are the flux values for the transient. All other values are the
+            flux values assigned to the points that make up the background model.
+        cutout_wcs_list : list
+            List of WCS objects for the cutouts used in the lightcurve analysis.
+        sim_lc : pd.DataFrame
+            The simulated lightcurve data, if applicable.
+        gal_ra : float
+            The RA of the host galaxy, if known.
+        gal_dec : float
+            The Dec of the host galaxy, if known.
+        """
+        self.diaobj = diaobj
+        self.flux = flux
+        self.sigma_flux = sigma_flux
+        self.images = images
+        self.model_images = model_images
+        self.image_list = image_list
+        self.cutout_image_list = cutout_image_list
+        self.ra_grid = ra_grid
+        self.dec_grid = dec_grid
+        self.wgt_matrix = wgt_matrix
+        self.confusion_metric = confusion_metric
+        self.best_fit_model_values = best_fit_model_values
+        self.sim_lc = sim_lc
+        self.galaxy_images = galaxy_images
+        self.noise_maps = noise_maps
+        self.galaxy_only_model_images = galaxy_only_model_images
+        self.galra = galra
+        self.galdec = galdec
+
+
 def gaussian(x, A, mu, sigma):
     """See name of function. :D"""
     return A * np.exp(-((x - mu) ** 2) / (2 * sigma**2))
